Add cpupossible package
