+++ pkg/cpurange/cpu_range.go
@@ -1,15 +1,12 @@
-package cpuonline
+package cpurange
 
 import (
-	"io/ioutil"
 	"strconv"
 	"strings"
 )
 
-const cpuOnline = "/sys/devices/system/cpu/online"
-
 // loosely based on https://github.com/iovisor/bcc/blob/v0.3.0/src/python/bcc/utils.py#L15
-func readCPURange(cpuRangeStr string) ([]uint, error) {
+func ReadCPURange(cpuRangeStr string) ([]uint, error) {
 	var cpus []uint
 	cpuRangeStr = strings.Trim(cpuRangeStr, "\n ")
 	for _, cpuRange := range strings.Split(cpuRangeStr, ",") {
@@ -31,13 +28,4 @@ func readCPURange(cpuRangeStr string) ([]uint, error) {
 		}
 	}
 	return cpus, nil
-}
-
-// Get returns a slice with the online CPUs, for example `[0, 2, 3]`
-func Get() ([]uint, error) {
-	buf, err := ioutil.ReadFile(cpuOnline)
-	if err != nil {
-		return nil, err
-	}
-	return readCPURange(string(buf))
 }
